SPO-13135 - поправлены замечания из пр: вращение лоадера, вынесена тема приложения в отдельный фай, поправены пэддинги между элементами
