+++ lib/features/list_place_screen/screens/places_list_screen/list_place_ui_elements/animate_big_loader.dart
@@ -1,3 +1,5 @@
+import 'dart:math' as math;
+
 import 'package:flutter/cupertino.dart';
 import 'package:test_case/assets/res/guide_icons.dart';
 
@@ -28,7 +30,7 @@ class _AnimateLoaderState extends State<AnimateLoader>
       animation: loaderSpinningController,
       builder: (context, child) {
         return Transform.rotate(
-          angle: loaderSpinningController.value * 25,
+          angle: loaderSpinningController.value * 2 * math.pi,
           child: child,
         );
       },
