+++ lib/features/list_place_screen/screens/places_list_screen/list_place_ui_elements/list_place_element.dart
@@ -16,7 +16,7 @@ class ListPlaceElement extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.all(16),
+      padding: const EdgeInsets.only(left: 16, right: 16, top: 24, bottom: 24),
       child: ClipRRect(
         borderRadius: BorderRadius.circular(10),
         child: ColoredBox(
